#215 Break out regex functino from TextHighlightRegex to helper file
